move provisional strategies to strings
